refactor: standalone swagger generation script
